requirements.txt
flask==1.1.2
requests==2.23.0
python-dotenv==0.15.0
ShopifyAPI==8.4.1
google-api-python-client==1.7.8
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
Flask-Mail==0.9.1
scikit-learn==0.24.2
pandas==1.1.5 
numpy==1.21.1 
Faker==8.10.3